chore: separate the helper functions
